corrupted packet counter done

+++ RF-UI-Interface/ProtocolDefine.h
@@ -40,7 +40,7 @@ namespace ui_interface {
         /// PacketNbr
         PACKET_RX_RATE_CTR, //uint32_t  **This is incremented on each packet received, eated by guito find the packet rate.
         RX_PACKET_CTR,//uint32_t        **This is the count of the received packets from the beginning of the program
-        CORRUPTED_PACKET_CTR,//uint32_t **The count of the corrupted packets received. Incremented by RX.
+        CORRUPTED_PACKET_CTR,//uint64_t **The count of the corrupted packets received. Incremented by RX.
 
         /// Header
         PACKET_ID, //uint8_t            **The ID of the last packet received
